Back/pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "ant-stock-backend"
version = "1.0.0"
description = "Backend moderno do ANT Stock."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.16.0",
    "asyncpg>=0.30.0",
    "celery[redis]>=5.5.0",
    "email-validator>=2.2.0",
    "fastapi[standard]>=0.115.0",
    "httpx>=0.28.0",
    "loguru>=0.7.3",
    "orjson>=3.10.18",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.0",
    "PyJWT>=2.10.1",
    "redis>=5.2.0",
    "sqlalchemy[asyncio]>=2.0.40",
    "sqlmodel>=0.0.24",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.11.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC", "S"]
ignore = ["B008", "S101", "S105", "UP046"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
strict = true
warn_unreachable = true
warn_unused_ignores = true
exclude = ["alembic/"]
disable_error_code = ["import-untyped"]
